move: components from app to assessments;
