+++ src/app/admin/assessments/assessments.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { TaskService } from '../tasks/task.service';
 import { Task } from '../tasks/task';
-import { QuestionService } from '../../services/question.service';
+import { QuestionService } from './question.service';
 
 @Component({
   selector: 'xp-assessments',
