+++ src/app/admin/assessments/open-answer-form/open-answer-form.component.ts
@@ -1,8 +1,8 @@
 import {Component, OnInit, Input, OnChanges} from '@angular/core';
-import {Question} from '../../../classes/question';
+import {Question} from '../question';
 import {Task} from '../../tasks/task';
 import {FormGroup, Validators, FormBuilder} from '@angular/forms';
-import {QuestionService} from '../../../services/question.service';
+import {QuestionService} from '../question.service';
 
 @Component({
   selector: 'xp-open-answer-form',
